Refactor brute
Instead of creating a DahuaController each pass
simply create it in BruteThread itself. In
theory, this should speed up the program as there
will be fewer calls to __init__ and creation or
destruction of objects.

Also, when the first error is caught during brute
process (mostly timeouts since the host can block
us or shut down at all), we stop it by return
statement.